CQRS implemented (bus not yet)

+++ Pinnacle.ResponsibleGaming.Domain/Configurations/SelfExclusionConfiguration.cs
@@ -20,7 +20,7 @@ namespace Pinnacle.ResponsibleGaming.Domain.Configurations
             //Properties
             Property(t => t.TimeInDays)
                 .HasColumnName("Limit")
-                .HasColumnType("int")
+                .HasColumnType("decimal")
                 .IsRequired();
         }
     }
